# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Tier Trade API
  version: 0.1.0
security:
  - bearerAuth: []
paths:
...
  /v1/finance/purchase-cost-components:
    post:
      summary: Registra um componente explícito da composição financeira da compra
      description: >-
        Exige FINANCE_EDIT. O componente pode aumentar, reduzir ou apenas documentar o valor
        a pagar; nenhuma taxa ou desconto é inferido pelo sistema.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreatePurchaseCostComponent' } } }
      responses:
        '201': { description: Componente persistido, auditado e refletido no saldo quando aplicável }
        '404': { description: Evento/título de compra não encontrado no tenant }
  /v1/finance/purchase-cost-components/{componentId}/reverse:
    post:
      summary: Estorna um componente de compra preservando o lançamento original
      parameters:
        - { name: componentId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ReversalReason' } } }
      responses:
        '201': { description: Componente e ajuste associados estornados }
        '409': { description: Componente já estornado }
  /v1/finance/policies:
    post:
      summary: Cria uma versão da política de alçada financeira
      description: Exige FINANCE_APPROVE e desativa a versão anterior sem apagá-la.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ConfigureFinancePolicy' } } }
      responses:
        '201': { description: Nova política ativa e versionada }
  /v1/finance/payment-batches:
    post:
      summary: Cria lote de pagamentos a partir de títulos a pagar
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreatePaymentBatch' } } }
      responses:
        '201': { description: Lote criado em rascunho }
        '409': { description: Um item não é título a pagar }
        '422': { description: Item superior ao saldo disponível }
  /v1/finance/payment-batches/{batchId}/submit:
    post:
      summary: Submete o lote à política financeira vigente
      parameters:
        - $ref: '#/components/parameters/PaymentBatchId'
      responses:
        '201': { description: Lote aprovado automaticamente ou enviado para aprovação }
        '409': { description: Política ausente ou lote fora do estado de rascunho }
  /v1/finance/payment-batches/{batchId}/approve:
    post:
      summary: Aprova lote acima da alçada automática com segregação de função
      parameters:
        - $ref: '#/components/parameters/PaymentBatchId'
      responses:
        '201': { description: Lote aprovado por usuário diferente do criador }
        '409': { description: Lote fora do estado ou violação da regra de quatro olhos }
  /v1/finance/payment-batches/{batchId}/execute:
    post:
      summary: Executa atomicamente os pagamentos de um lote aprovado
      parameters:
        - $ref: '#/components/parameters/PaymentBatchId'
      responses:
        '201': { description: Pagamentos gerados e lote marcado como executado }
        '409': { description: Lote ainda não aprovado }
  /v1/finance/bank-accounts:
    post:
      summary: Cadastra conta bancária operacional do tenant
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateBankAccount' } } }
      responses:
        '201': { description: Conta bancária cadastrada }
  /v1/finance/bank-statement-entries:
    post:
      summary: Importa manualmente uma linha de extrato para conciliação
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateBankStatementEntry' } } }
      responses:
        '201': { description: Linha registrada como não conciliada }
  /v1/finance/bank-statement-entries/{entryId}/reconcile:
    post:
      summary: Concilia uma linha de extrato com recebimento ou pagamento de mesmo valor e direção
      parameters:
        - { name: entryId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ReconcileBankStatementEntry' } } }
      responses:
        '201': { description: Linha conciliada com vínculo auditável }
        '409': { description: Linha já conciliada }
        '422': { description: Valor ou direção incompatível }
  /v1/finance/titles:
    post:
      summary: Emite título a receber para uma previsão financeira pronta
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateFinancialTitle' } } }
      responses:
        '201': { description: Título emitido }
        '409': { description: Título existente ou política de arredondamento pendente }
  /v1/finance/titles/{titleId}/settlements:
    post:
      summary: Registra recebimento parcial ou total de um título
      parameters:
        - { name: titleId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SettleFinancialTitle' } } }
      responses:
        '201': { description: Recebimento registrado e saldo atualizado }
        '422': { description: Valor superior ao saldo do título }
  /v1/finance/settlements/{settlementId}/reverse:
    post:
      summary: Estorna uma baixa preservando o registro original
      parameters:
        - { name: settlementId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reason]
              properties: { reason: { type: string, minLength: 3, maxLength: 500 } }
      responses:
        '201': { description: Baixa estornada e saldo recomposto }
        '409': { description: Baixa já estornada }
  /v1/finance/titles/{titleId}/payments:
    post:
      summary: Registra pagamento parcial ou total de um título fiscal
      parameters:
        - { name: titleId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PayFinancialTitle' } } }
      responses:
        '201': { description: Pagamento registrado; título, obrigação e caixa atualizados atomicamente }
        '409': { description: Título não é uma obrigação fiscal ou referência bancária já utilizada }
        '422': { description: Pagamento superior ao saldo em aberto }
  /v1/finance/payments/{paymentId}/reverse:
    post:
      summary: Estorna pagamento fiscal preservando o registro original
      parameters:
        - { name: paymentId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reason]
              properties: { reason: { type: string, minLength: 3, maxLength: 500 } }
      responses:
        '201': { description: Pagamento estornado; saldos e status recompostos }
        '409': { description: Pagamento já estornado }
...
  /v1/finance/commission-policies:
    post:
      summary: Publica versão de política de comissão configurável
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateCommissionPolicy' } } }
      responses: { '201': { description: Política versionada } }
  /v1/finance/commission-accruals:
    post:
      summary: Apropria comissão sobre evento financeiro auditado
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AccrueCommission' } } }
      responses:
        '201': { description: Comissão apropriada }
        '422': { description: Política de arredondamento necessária }
  /v1/documents:
    get:
      summary: Lista metadados e assinaturas de documentos do tenant
      parameters:
        - { name: aggregateType, in: query, schema: { type: string } }
        - { name: aggregateId, in: query, schema: { type: string, format: uuid } }
      responses: { '200': { description: Documentos visíveis ao usuário autenticado } }
  /v1/documents/upload-request:
    post:
      summary: Cria metadados e URL assinada para upload privado
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateDocumentUpload' } } }
      responses: { '201': { description: Upload privado autorizado por tempo limitado } }
  /v1/documents/{documentId}/complete:
    post:
      summary: Confirma que o arquivo foi recebido no bucket privado
      parameters: [{ $ref: '#/components/parameters/DocumentId' }]
      responses: { '201': { description: Documento disponível } }
  /v1/documents/{documentId}/download:
    get:
      summary: Gera URL temporária de download após validar tenant e membership
      parameters: [{ $ref: '#/components/parameters/DocumentId' }]
      responses: { '200': { description: URL privada válida por 60 segundos } }
  /v1/documents/{documentId}/signatures:
    post:
      summary: Registra estado de assinatura do documento
      parameters: [{ $ref: '#/components/parameters/DocumentId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RecordDocumentSignature' } } }
      responses: { '201': { description: Assinatura registrada } }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    PaymentBatchId:
      name: batchId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    DocumentId:
      name: documentId
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    CreateCommissionPolicy:
      type: object
      additionalProperties: false
      required: [code, name, basis, ratePct, beneficiaryName, effectiveFrom, status]
      properties:
        code: { type: string, pattern: '^[A-Z0-9][A-Z0-9_-]{1,39}$' }
        name: { type: string, minLength: 3, maxLength: 120 }
        basis: { enum: [FINANCIAL_EVENT_AMOUNT] }
        ratePct: { type: string, pattern: '^\d+(?:\.\d{1,6})?$' }
        commodity: { enum: [MILHO, SOJA, null] }
        beneficiaryName: { type: string, minLength: 2, maxLength: 160 }
        effectiveFrom: { type: string, format: date }
        effectiveTo: { type: [string, 'null'], format: date }
        status: { enum: [DRAFT, ACTIVE] }
    AccrueCommission:
      type: object
      additionalProperties: false
      required: [policyId, financialEventId]
      properties:
        policyId: { type: string, format: uuid }
        financialEventId: { type: string, format: uuid }
    CreateDocumentUpload:
      type: object
      additionalProperties: false
      required: [aggregateType, aggregateId, documentType, fileName, mimeType, sizeBytes]
      properties:
        aggregateType: { enum: [CONTRACT, SALES_CONTRACT, LOAD, FISCAL_DOCUMENT, COUNTERPARTY, INVENTORY_LOT] }
        aggregateId: { type: string, format: uuid }
        documentType: { enum: [CONTRACT_DRAFT, SIGNED_CONTRACT, AMENDMENT, GUARANTEE, INVOICE, ROMANEIO, QUALITY_REPORT, WEIGHING_TICKET, OTHER] }
        fileName: { type: string, minLength: 1, maxLength: 240 }
        mimeType: { enum: [application/pdf, image/jpeg, image/png, application/vnd.openxmlformats-officedocument.wordprocessingml.document, application/vnd.openxmlformats-officedocument.spreadsheetml.sheet] }
        sizeBytes: { type: integer, minimum: 1, maximum: 26214400 }
        notes: { type: [string, 'null'], minLength: 1, maxLength: 1000 }
    RecordDocumentSignature:
      type: object
      additionalProperties: false
      required: [provider, signerName, signerRole, status]
      properties:
        provider: { enum: [MANUAL, DOCUSIGN, OTHER] }
        externalEnvelopeId: { type: [string, 'null'] }
        signerName: { type: string, minLength: 2, maxLength: 160 }
        signerEmail: { type: [string, 'null'], format: email }
        signerRole: { type: string, minLength: 2, maxLength: 80 }
        status: { enum: [PENDING, SENT, SIGNED, DECLINED, CANCELLED] }
        sentAt: { type: [string, 'null'], format: date-time }
        signedAt: { type: [string, 'null'], format: date-time }
...
    CreateFinancialTitle:
      type: object
      additionalProperties: false
      required: [financialEventId, titleNumber, documentReference, dueDate]
      properties:
        financialEventId: { type: string, format: uuid }
        titleNumber: { type: string, minLength: 3, maxLength: 40 }
        documentReference: { type: string, minLength: 1, maxLength: 80 }
        dueDate: { type: string, format: date }
    SettleFinancialTitle:
      type: object
      additionalProperties: false
      required: [amount, receivedAt, bankReference, notes]
      properties:
        amount: { type: string, pattern: '^\d+(?:\.\d{1,2})?$' }
        receivedAt: { type: string, format: date-time }
        bankReference: { type: string, minLength: 1, maxLength: 80 }
        notes: { type: [string, 'null'], maxLength: 1000 }
    PayFinancialTitle:
      type: object
      additionalProperties: false
      required: [amount, paidAt, bankReference, notes]
      properties:
        amount: { type: string, pattern: '^\d+(?:\.\d{1,2})?$' }
        paidAt: { type: string, format: date-time }
        bankReference: { type: string, minLength: 1, maxLength: 80 }
        notes: { type: [string, 'null'], maxLength: 1000 }
    ReversalReason:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason: { type: string, minLength: 3, maxLength: 500 }
    CreatePurchaseCostComponent:
      type: object
      additionalProperties: false
      required: [financialEventId, componentType, payableImpact, amount, description, externalReference]
      properties:
        financialEventId: { type: string, format: uuid }
        componentType: { type: string, enum: [QUALITY_DISCOUNT, FREIGHT, STORAGE, TAX_WITHHOLDING, OTHER] }
        payableImpact: { type: string, enum: [REDUCE_PAYABLE, INCREASE_PAYABLE, MEMO_ONLY] }
        amount: { type: string, pattern: '^\d+(?:\.\d{1,2})?$' }
        description: { type: string, minLength: 3, maxLength: 500 }
        externalReference: { type: [string, 'null'], minLength: 1, maxLength: 80 }
    ConfigureFinancePolicy:
      type: object
      additionalProperties: false
      required: [paymentApprovalThreshold]
      properties:
        paymentApprovalThreshold: { type: string, pattern: '^\d+(?:\.\d{1,2})?$' }
    CreatePaymentBatch:
      type: object
      additionalProperties: false
      required: [reference, scheduledOn, items]
      properties:
        reference: { type: string, minLength: 3, maxLength: 40 }
        scheduledOn: { type: string, format: date }
        items:
          type: array
          minItems: 1
          maxItems: 200
          items:
            type: object
            additionalProperties: false
            required: [titleId, amount]
            properties:
              titleId: { type: string, format: uuid }
              amount: { type: string, pattern: '^\d+(?:\.\d{1,2})?$' }
    CreateBankAccount:
      type: object
      additionalProperties: false
      required: [code, name]
      properties:
        code: { type: string, pattern: '^[A-Z0-9][A-Z0-9_-]{1,31}$' }
        name: { type: string, minLength: 3, maxLength: 120 }
    CreateBankStatementEntry:
      type: object
      additionalProperties: false
      required: [bankAccountId, occurredAt, direction, amount, bankReference, description]
      properties:
        bankAccountId: { type: string, format: uuid }
        occurredAt: { type: string, format: date-time }
        direction: { type: string, enum: [CREDIT, DEBIT] }
        amount: { type: string, pattern: '^\d+(?:\.\d{1,2})?$' }
        bankReference: { type: string, minLength: 1, maxLength: 80 }
        description: { type: [string, 'null'], minLength: 1, maxLength: 500 }
    ReconcileBankStatementEntry:
      type: object
      additionalProperties: false
      required: [matchedType, matchedId]
      properties:
        matchedType: { type: string, enum: [SETTLEMENT, PAYMENT] }
        matchedId: { type: string, format: uuid }
...